feat(wasm): basic integration for wasm
