+++ russh-util/Cargo.toml
@@ -6,21 +6,15 @@ rust-version = "1.65"
 
 [dependencies]
 chrono = "0.4.38"
+tokio = { version = "1.17", features = ["sync", "macros"] }
 
 [dev-dependencies]
 futures-executor = "0.3.13"
 static_assertions = "1.1.0"
-
 
 [target.'cfg(target_arch = "wasm32")'.dependencies]
 wasm-bindgen = "0.2"
 wasm-bindgen-futures = "0.4.43"
 
 [target.'cfg(not(target_arch = "wasm32"))'.dependencies]
-tokio = { version = "1.17", features = [
-    "io-util",
-    "macros",
-    "sync",
-    "rt-multi-thread",
-    "rt",
-] }
+tokio = { version = "1.17", features = ["io-util", "rt-multi-thread", "rt"] }
