+++ russh-util/src/lib.rs
@@ -1,5 +1,2 @@
-#![allow(dead_code)] // To be removed when full wasm support is added.
-
-pub mod future;
 pub mod runtime;
 pub mod time;
