+++ russh/Cargo.toml
@@ -38,7 +38,6 @@ hmac = { workspace = true }
 log = { workspace = true }
 num-bigint = { version = "0.4", features = ["rand"] }
 once_cell = "1.13"
-openssl = { workspace = true, optional = true }
 p256 = { version = "0.13", features = ["ecdh"] }
 p384 = { version = "0.13", features = ["ecdh"] }
 p521 = { version = "0.13", features = ["ecdh"] }
@@ -53,16 +52,9 @@ ssh-encoding = { workspace = true }
 ssh-key = { workspace = true }
 subtle = "2.4"
 thiserror = { workspace = true }
-tokio = { workspace = true, features = [
-    "io-util",
-    "rt-multi-thread",
-    "time",
-    "net",
-    "sync",
-    "macros",
-    "process",
-] }
+russh-util = { path = "../russh-util" }
 des = "0.8.1"
+tokio = { workspace = true, features = ["io-util", "sync", "time"] }
 
 [dev-dependencies]
 anyhow = "1.0"
@@ -77,7 +69,6 @@ tokio = { version = "1.17.0", features = [
     "sync",
     "macros",
 ] }
-russh-sftp = "2.0.0-beta.2"
 rand = "0.8.5"
 shell-escape = "0.1"
 tokio-fd = "0.3"
@@ -86,3 +77,8 @@ ratatui = "0.26.0"
 
 [package.metadata.docs.rs]
 features = ["openssl"]
+
+[target.'cfg(not(target_arch = "wasm32"))'.dependencies]
+openssl = { workspace = true, optional = true }
+russh-sftp = "2.0.0-beta.2"
+tokio = { workspace = true }
